Moved the SavedQueryManager class from PCAxis.Query to PxWeb. Removed system.runtime.caching from PCAxis.Query.
