+++ PCAxis.Query/PCAxis.Query.csproj
@@ -90,7 +90,6 @@
     <Reference Include="System.Data.OracleClient" />
     <Reference Include="System.Drawing" />
     <Reference Include="System.Net" />
-    <Reference Include="System.Runtime.Caching" />
     <Reference Include="System.Security" />
     <Reference Include="System.Security.AccessControl, Version=4.1.3.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a, processorArchitecture=MSIL">
       <HintPath>..\packages\System.Security.AccessControl.4.7.0\lib\net461\System.Security.AccessControl.dll</HintPath>
@@ -132,7 +131,6 @@
     <Compile Include="OperationsTracker.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="QueryHelper.cs" />
-    <Compile Include="SavedQueryManager.cs" />
     <Compile Include="Serializers\CalculatePerPartSerializer.cs" />
     <Compile Include="Serializers\ChangeDecimalsSerializer.cs" />
     <Compile Include="Serializers\ChangeTextCodePresentationSerializer.cs" />
