+++ PXWeb/Admin/Features-SaveQuery-General.aspx.cs
@@ -73,9 +73,9 @@ namespace PXWeb.Admin
                     {
                         PXWeb.Settings.EndUpdate();
 
-                        PCAxis.Query.SavedQueryManager.StorageType = PXWeb.Settings.Current.Features.SavedQuery.StorageType;
-                        
-                        PCAxis.Query.SavedQueryManager.Reset();
+                        PXWeb.Management.SavedQueryManager.StorageType = PXWeb.Settings.Current.Features.SavedQuery.StorageType;
+
+                        PXWeb.Management.SavedQueryManager.Reset();
                         PXWeb.Management.SavedQueryPaxiomCache.Current.Reset();
                     }
                 }
