+++ PXWeb/Code/Management/SavedQueryManager.cs
@@ -1,14 +1,13 @@
-﻿using PCAxis.Query;
+﻿using Newtonsoft.Json;
+using PCAxis.Query;
 using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
-using System.Web;
-using Newtonsoft.Json;
-using System.Web.Hosting;
 using System.Runtime.Caching;
+using System.Web;
 
-namespace PCAxis.Query
+namespace PXWeb.Management
 {
     public class SavedQueryManager
     {
